Updated routes, fixing edit functionality, updated edit_mutation and company_by_id query
