+++ bio-bid/src/App.js
@@ -3,10 +3,9 @@ import { Route } from "react-router-dom";
 
 // Component Imports
 import Bids from './components/bids';
-import Form from './components/company-profile/Form/Form';
 import Details from './components/company-profile/Details/Details';
 import List from './components/company-profile/List/List';
-import NewForm from './components/company-profile/NewForm/Form';
+import Form from './components/company-profile/Form/Form';
 
 import Dashboard from './components/admin/Dashboard/Dashboard';
 
@@ -19,20 +18,17 @@ function App() {
       <Route exact path='/'>
         <List/>
       </Route>
-      {/* <Route path='/service-provider/add'>
+      <Route path='/service-provider/add'>
         <Form edit={false} />
-      </Route> */}
+      </Route>
       <Route path='/service-providers/:id'>
         <Details />
       </Route>
-      {/* <Route path='/service-provider/edit/:id'>
+      <Route path='/service-provider/edit/:id'>
         <Form edit={true} />
-      </Route> */}
+      </Route>
       <Route path='/admin/dashboard'>
         <Dashboard/>
-      </Route>
-      <Route path='/dev-form'>
-        <NewForm/>
       </Route>
     </div>
   );
