+++ bio-bid/src/components/company-profile/Form/Form.js
@@ -1,542 +1,730 @@
-// import React, { useState, useEffect } from 'react';
-// import { useQuery, useMutation } from '@apollo/react-hooks';
-// import { GET_COMPANY_BY_ID, GET_REGIONS, GET_THERAPEUTICS, GET_SERVICES, GET_SPECIALTIES } from '../../../queries';
-// import { ADD_COMPANY, EDIT_COMPANY } from '../../../mutations';
-// import { useHistory, useParams } from 'react-router-dom';
-// import { makeStyles } from '@material-ui/core/styles';
+import React, { useState, useEffect } from 'react';
+import { useQuery, useMutation } from '@apollo/react-hooks';
+import { useHistory, useParams } from 'react-router-dom';
 
-// import WarningCard from './WarningCard';
-// import MultipleInput from './MultipleInput';
-// import Backdrop from '@material-ui/core/Backdrop';
-// import CircularProgress from '@material-ui/core/CircularProgress';
-// import { Body, LinkedIn, Button, Form, HeaderError, SucceessMessage } from './styles';
+import WarningCard from './WarningCard';
+import Service from './Service';
 
-// import scienceAsset from '../../../images/science-asset.svg';
-// import defaultLogo from '../../../images/default-company-logo.png';
-// import theme from '../../../theme';
+import { GET_REGIONS, GET_THERAPEUTICS, GET_SERVICES, GET_SPECIALTIES, GET_COMPANY_BY_ID } from '../../../queries';
+import { ADD_COMPANY, EDIT_COMPANY } from '../../../mutations';
 
-// const useStyles = makeStyles((theme) => ({
-//     backdrop: {
-//       zIndex: theme.zIndex.drawer + 1,
-//       color: '#fff',
-//     },
-//   }));
+import Backdrop from '@material-ui/core/Backdrop';
+import CircularProgress from '@material-ui/core/CircularProgress';
+import { makeStyles } from '@material-ui/core/styles';
+import { Body, Basic, Services, Button, WarningIcon, Arrow } from './styles';
 
-// export default (props) => {
-//     // id will only be available if props.edit === true
-//     const { id } = useParams();
+import Input from '@material-ui/core/Input';
+import MenuItem from '@material-ui/core/MenuItem';
+import FormControl from '@material-ui/core/FormControl';
+import ListItemText from '@material-ui/core/ListItemText';
+import Select from '@material-ui/core/Select';
+import Checkbox from '@material-ui/core/Checkbox';
 
-//     const classes = useStyles();
-//     const history = useHistory();
+const useStyles = makeStyles((theme) => ({
+    backdrop: {
+      zIndex: theme.zIndex.drawer + 1,
+      color: '#fff',
+    },
+    formControl: {
+        minWidth: 120,
+        maxWidth: 300,
+    },
+        chips: {
+        display: 'flex',
+        flexWrap: 'wrap',
+    },
+    noLabel: {
+        marginTop: theme.spacing(3),
+    },
+}));
 
-//     /*
-//         open - Boolean state for whether the backdrop for warnings/loading should be opened.
-//         importWarning - Boolean state for warning type when attempting to import
-//         linkedInError - Boolean state based on if the linkedInUrl in the `linkedin` input field is valid
-//         appError - Stores any errors that exist in the Form component
-//         headerError - Boolean state for whether an error should be displayed or not
-//         headerErrorText - Stores the value of the error
-//         headerSuccess- Boolean state for when changes are saved successfully
-//         logo - The current logo image pulled in from LinkedIn/default
-//         regions, therapeutics, services, specialties - State to keep track of the mapped 
-//             data from the database for these properties
-//         newId - When creating a new company profile, this is the new id is stored here
-//         addLoading - State to determine if the mutation to add/update a company is loading
-//     */
-//     const [ open, setOpen ] = useState(false);
-//     const [ importWarning, setImportWarning ] = useState(false);
-//     const [ linkedInError, setLinkedInError] = useState(false);
-//     const [ appError, setAppError] = useState(null);
-//     const [ headerError, setHeaderError ] = useState(false);
-//     const [ headerErrorText, setHeaderErrorText ] = useState('');
-//     const [ headerSuccess, setHeaderSuccess ] = useState(false);
-//     const [ logo, setLogo ] = useState(defaultLogo);
-//     const [ regions, setRegions ] = useState(null);
-//     const [ therapeutics, setTherapeutics ] = useState(null);
-//     const [ services, setServices ] = useState(null);
-//     const [ specialties, setSpecialties ] = useState(null);
-//     const [ newId, setNewId ] = useState(null);
-//     const [ addLoading, setAddLoading ] = useState(false);
+const ITEM_HEIGHT = 48;
+const ITEM_PADDING_TOP = 8;
+const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250,
+    },
+  },
+};
 
-//     /*
-//         useQuery and useMutation
-//         - useQuery will accept a gql query that is defined in the `src/queries` directory.
-//             - Conditionally returns a `loading` boolean true/false based on if the query is running
-//             - Conditionally returns an `error` object if there was an error with the query
-//             - Conditionally returns the `data` requested by the query if successful
-//         - useMutation will accept a gql mutation that is defined in the `src/mutations` directory.
-//             - Returns a functions that can be used to insert data into the mutation as arguments
-//             - Example: 
-//                 addCompany({ variables: formData }) 
-//                 addCompany({ variables: {name: formData.name, imgURL: formData.imgURL} })
-            
-//     */
-//     const { loading: companyLoading, error: companyError, data: companyData } = useQuery(GET_COMPANY_BY_ID, { 
-//         variables: { id },
-//         skip: !props.edit
-//     });
+const phaseOptions = [
+    'I',
+    'II',
+    'III',
+    'IV'
+];
 
-//     const { error: regionsError, data: regionsData } = useQuery(GET_REGIONS);
-//     const { error: therapeuticsError, data: therapeuticsData } = useQuery(GET_THERAPEUTICS);
-//     const { error: servicesError, data: servicesData} = useQuery(GET_SERVICES);
-//     const { error: specialtiesError, data: specialtiesData } = useQuery(GET_SPECIALTIES);
+export default ({edit}) => {
+    const classes = useStyles();
+    const history = useHistory();
+    const { id } = useParams();
 
-//     const [ addCompany ] = useMutation(ADD_COMPANY, {
-//         update: (proxy, result) => {
-//             setNewId(result.data.createCompany.id);
-//         }
-//     });
+    // Queries
+    const { data: regionsData } = useQuery(GET_REGIONS);
+    const { data: therapeuticsData } = useQuery(GET_THERAPEUTICS);
+    const { data: servicesData } = useQuery(GET_SERVICES);
+    const { data: specialtyData } = useQuery(GET_SPECIALTIES);
+    const { loading, error: companyError, data: companyData } = useQuery(GET_COMPANY_BY_ID, {
+        variables: { id },
+        skip: !edit
+    })
 
-//     const [ editCompany ] = useMutation(EDIT_COMPANY);
+    // Mutations
+    const [ addCompany ] = useMutation(ADD_COMPANY, {
+        update: (proxy, result) => {
+            setNewId(result.data.createCompany.id);
+        }
+    });
+    const [ editCompany ] = useMutation(EDIT_COMPANY);
 
-//     /*
-//         formData
-//         - Inputs in the form components are controlled with `handleUpdate` and `handleMultiUpdate`.
-//         - Everything is saved in formData state until the component is unMounted 
-//             or there is a successful submission in `handleSubmit`
-//     */
-//     const [ formData, setFormData ] = useState({
-//         name: '',
-//         logoURL: '',
-//         website: '',
-//         linkedin: '',
-//         overview: '',
-//         headquarters: '',
-//         companySize: '',
-//         services: [],
-//         specialties: [],
-//         regionsCovered: [],
-//         therapeuticAreas: []
-//     })
+    // State
+    const [ confirmCancel, setConfirmCancel ] = useState(false);
+    const [ formData, setFormData ] = useState({
+        name: '',
+        logoURL: '',
+        website: '',
+        email: '',
+        linkedin: '',
+        overview: '',
+        headquarters: '',
+        companySize: ''
+    })
 
-//     /*
-//         handleImportWarning
-//         - Checks to se if there is a valid LinkedInURL. The state of `linkedInError` is set
-//             in a useEffect hook that checks the linkedin input every time it is updated
-//     */
-//     const handleImportWarning = () => {
-//         if(!linkedInError){
-//             setImportWarning(true);
-//             setOpen(true);
-//         }
-//     }
+    const [ phases, setPhases ] = useState([]);
+    const [ regions, setRegions ] = useState([]);
+    const [ therapeutics, setTherapeutics ] = useState([]);
+    const [ services, setServices ] = useState([]);
 
-//     const handleCancelWarning = () => {
-//         setOpen(true);
-//     }  
+    const [ regionsAll, setRegionsAll ] = useState(false);
+    const [ therapeuticsAll, setTherapeuticsAll ] = useState(false);
+    const [ phasesAll, setPhasesAll ] = useState(false);
 
-//     const handleWarningClose = () => {
-//         setImportWarning(false);
-//         setOpen(false);
-//     }
+    const [ error, setError ] = useState('');
+    const [ newId, setNewId ] = useState(null);
+    const [ submitting, setSubmitting ] = useState(false);
 
-//     const handleUpdate = e => {
-//         setFormData({
-//             ...formData,
-//             [e.target.name]: e.target.value
-//         })
-//     }
+    // Event handlers
+    const handleMultiple = e => {
+        if(e.target.name === 'regions'){
+            setRegions(e.target.value);
+        }
+        if(e.target.name === 'therapeutics'){
+            setTherapeutics(e.target.value);
+        }   
+        if(e.target.name === 'phases'){
+            setPhases(e.target.value);
+        }
+        if(e.target.name === 'services'){
+            /*
+                services state = old array
+                e.target.value = new array
+                Compare [new and old array] to see if a services has been removed or added
+                    - If a services has been added, add the new service to [old array]
+                    - If a services has been removed, remove the service from [old array]
+                Now that [old array] has been changed, it has updated state
 
-//     const handleMultiUpdate = (name, element) => {
-//         setFormData({
-//             ...formData,
-//             [name]: element.map(value => {
-//                 return {'name': value}
-//             })
-//         })
-//     }
+                Comparing Algorithm
+                    - Check the length of [old array]
+                        - If [new array] is longer, we know there is a new service.
+                            - Use [new array] and pull the last index and add it to [old array]
+                        - If [new array] is shorter, we know a service has been removed.
+                            - Alphabetically sort [old array and new array]
+                            - Starting at index 0, compare each index of both arrays
+                                - When a value doesn't match in the old array, this is the removed service
+                                    - Remove that service from the original services state
+            */
+            if(e.target.value.length > services.length){
+                const newService = e.target.value[e.target.value.length - 1];
+                setServices([
+                    ...services,
+                    {
+                        name: newService,
+                        specialties: []
+                    }
+                ])
+            }else{
+                const oldArray = services.map(service => service.name);
+                const newArray = e.target.value;
+                oldArray.sort();
+                newArray.sort();
+                let i=0;
+                while(true){
+                    if(oldArray[i] === newArray[i]){
+                        i++;
+                    }else{
+                        setServices(services.filter(service => service.name !== oldArray[i]));
+                        return;
+                    }
+                }
+            }
 
-//     /*
-//         Submit event handler
-//         - Handles sending the formData to backend for adding a new company
-//         - Handles sending the formData to the backend for edits
-//         - Checks to make sure that the name field isn't null or empty
-//         - Handles loading state to display a spinner while mutations run
-//     */
-//     const handleSubmit = async(e) => {
-//         e.preventDefault();
-//         if(props.edit){
-//             if(!formData.name){
-//                 setHeaderError(true);
-//                 setHeaderErrorText('Missing required field: Company Name');
-//                 window.scrollTo(0, 0);
-//             }else{
-//                 setHeaderError(false);
-//                 setHeaderErrorText('');
-//                 setAddLoading(true);
-//                 try{
-//                     await editCompany({ variables: {
-//                         id: id,
-//                         name: formData.name,
-//                         logoURL: formData.logoURL,
-//                         website: formData.website,
-//                         linkedin: formData.linkedin,
-//                         overview: formData.overview,
-//                         headquarters: formData.headquarters,
-//                         companySize: formData.companySize === '' ? null : formData.companySize,
-//                         regions: formData.regionsCovered,
-//                         therapeutics: formData.therapeuticAreas,
-//                         services: formData.services,
-//                         specialties: formData.specialties
-//                     }});
-//                     setAddLoading(false);
-//                     setHeaderSuccess(true);
-//                 }
-//                 catch(error){
-//                     console.log(error);
-//                     setAddLoading(false);
-//                     setHeaderError(true);
-//                     setHeaderErrorText('The company name already exists');
-//                 }
-//             }
-//         }else{
-//             if(!formData.name){
-//                 setHeaderError(true);
-//                 setHeaderErrorText('Missing required field: Company Name');
-//                 window.scrollTo(0, 0);
-//             }else{
-//                 setHeaderError(false);
-//                 setHeaderErrorText('');
-//                 setAddLoading(true);
-//                 try{
-//                     await addCompany({ variables: {
-//                         name: formData.name,
-//                         logoURL: formData.logoURL,
-//                         website: formData.website,
-//                         linkedin: formData.linkedin,
-//                         overview: formData.overview,
-//                         headquarters: formData.headquarters,
-//                         companySize: formData.companySize === '' ? null : formData.companySize,
-//                         regions: formData.regionsCovered,
-//                         therapeutics: formData.therapeuticAreas,
-//                         services: formData.services,
-//                         specialties: formData.specialties
-//                     }});
-//                     setAddLoading(false);
-//                 }
-//                 catch(error){
-//                     console.log(error);
-//                     setAddLoading(false);
-//                     setHeaderError(true);
-//                     setHeaderErrorText('The company name already exists');
-//                 }
-//             }
-//         }        
-//     }
+        }
+    }
 
-//     const handleReDirect = () => {
-//         if(props.edit){
-//             history.push(`/service-providers/${id}`);
-//         }else{
-//             history.push('/')
-//         }
-//     }
+    const setAll = field => {
+        if(field === 'regions'){
+            setRegions(regionsData.regions.map(region => region.name));
+        }
+        if(field === 'therapeutics'){
+            setTherapeutics(therapeuticsData.therapeutics.map(therapeutic => therapeutic.name));
+        }
+        if(field === 'phases'){
+            setPhases(phaseOptions);
+        }
+    }
 
-//     const validateUrl = url => {
-//         const regex = /^(?:(?:https?|ftp):\/\/)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:\/\S*)?$/;
-//         if(regex.test(url)) return true;
-//         return false;
-//     }
+    const clearAll = field => {
+        if(field === 'regions'){
+            setRegions([]);
+        }
+        if(field === 'therapeutics'){
+            setTherapeutics([]);
+        }
+        if(field === 'phases'){
+            setPhases([])
+        }
+    }
 
-//     /*
-//         Handle LinkedIn error
-//         - If the url in the linkedin input field is invalid, display an error below the input field
-//         - If the url is valid, the error goes away
-//     */
-//     useEffect(() => {
-//         if(formData.linkedin && validateUrl(formData.linkedin)){
-//             setLinkedInError(false);
-//             setAppError('');
-//         }else{
-//             if(formData.linkedin.length > 0){
-//                 setLinkedInError(true);
-//                 setAppError('Must enter a valid LinkedIn URL')
-//             }else{
-//                 setLinkedInError(false);
-//                 setAppError('');
-//             } 
-//         }
-//     }, [ formData.linkedin ])
+    const toggleAllTherapeutics = () => {
+        setTherapeuticsAll(!therapeuticsAll);
+    }
 
-//     /*
-//         Fill in current data
-//         - When editing a company profile, chances are that the profile is already filled out
-//         - All current data stored with the profile in the database is filled in here to
-//             display on the UI for the user
-//     */
-//     useEffect(() => {
-//         if(props.edit){
-//             if(companyData){
-//                 setHeaderErrorText('');
-//                 setHeaderError(false);
-//                 const regionsMapped = companyData.company.regions.map(region => ({ name: region.name }));
-//                 const therapeuticsMapped = companyData.company.therapeutics.map(therapeutic => ({ name: therapeutic.name }));
-//                 const servicesMapped = companyData.company.services.map(service => ({ name: service.name }));
-//                 const specialtiesMapped = companyData.company.specialties.map(specialty => ({ name: specialty.name }));
-//                 setFormData({
-//                     name: companyData.company.name,
-//                     logoURL: companyData.company.logoURL,
-//                     website: companyData.company.website,
-//                     linkedin: companyData.company.linkedin,
-//                     overview: companyData.company.overview,
-//                     headquarters: companyData.company.headquarters,
-//                     companySize: companyData.company.companySize ? companyData.company.companySize : '',
-//                     regionsCovered: regionsMapped,
-//                     therapeuticAreas: therapeuticsMapped,
-//                     services: servicesMapped,
-//                     specialties: specialtiesMapped
-//                 })
-//             }
-//             if(companyError){
-//                 setHeaderError(true);
-//                 setHeaderErrorText('Error fetching previous company data');
-//             }
-//         }
-//     }, [ props.edit, companyData, companyError ])
+    const toggleAllRegions = () => {
+        setRegionsAll(!regionsAll);
+    }
 
-//     /*
-//         Updating suggestive data
-//         - Suggestive data is querried from backend and stored in their respective array
-//         - This useEffects updates and structures the state so that it may be passed 
-//             to its appropiate MultipleInput component
-//         - TODO?: Display any errors in the UI for feedback to the user
-//     */
-//     useEffect(() => {
-//         if(regionsData){
-//             const mappedData = regionsData.regions.map(region => ({ name: region.name }))
-//             setRegions(mappedData);
-//         }
-//         if(regionsError){
-//             console.log(regionsError);
-//         }
+    const toggleAllPhases = () => {
+        setPhasesAll(!phasesAll);
+    }
 
-//         if(therapeuticsData){
-//             const mappedData = therapeuticsData.therapeutics.map(therapeutic => ({ name: therapeutic.name }));
-//             setTherapeutics(mappedData);
-//         }
-//         if(therapeuticsError){
-//             console.log(therapeuticsError);
-//         }
+    const handleCancel = () => {
+        setConfirmCancel(!confirmCancel);
+    }
 
-//         if(servicesData){
-//             const mappedData = servicesData.services.map(service => ({ name: service.name }));
-//             setServices(mappedData);
-//         }
-//         if(servicesError){
-//             console.log(servicesError);
-//         }
+    const handleReDirect = () => {
+        history.push('/');
+    }
 
-//         if(specialtiesData){
-//             const mappedData = specialtiesData.specialties.map(specialty => ({ name: specialty.name }));
-//             setSpecialties(mappedData);
-//         }
-//         if(specialtiesError){
-//             console.log(specialtiesError);
-//         }
-//     }, [ regionsData, regionsError, therapeuticsData, therapeuticsError, servicesData, servicesError, specialtiesData, specialtiesError ]);
+    const handleSubmit = async () => {
+        if(!formData.name){
+            setError('Missing name field');
+            window.scrollTo(0, 0);
+        }else{
+            setSubmitting(true);
+            if(edit){
+                try{
+                    console.log({
+                        ...formData,
+                        regions: regions.map(region => ({ name: region })),
+                        therapeutics: therapeutics.map(therapeutic => ({ name: therapeutic })),
+                        phases: phases.length === 0 ? [] : phases
+                    })
+                    await editCompany({ variables: {
+                        id: id,
+                        name: formData.name,
+                        email: formData.email,
+                        logoURL: formData.logoURL,
+                        website: formData.website,
+                        linkedin: formData.linkedin,
+                        overview: formData.overview,
+                        headquarters: formData.headquarters,
+                        companySize: formData.companySize === '' ? null : formData.companySize,
+                        regions: regions.map(region => ({ name: region })),
+                        therapeutics: therapeutics.map(therapeutic => ({ name: therapeutic })),
+                        phases: phases.length === 0 ? [] : phases,
+                        services: services.length === 0 ? [] : services
+                    }})
+                    setSubmitting(false);
+                }catch(error){
+                    console.log('Save changes error ', error);
+                    setSubmitting(false);
+                }
+            }else{
+                try{
+                    await addCompany({ variables: {
+                        name: formData.name,
+                        logoURL: formData.logoURL,
+                        email: formData.email,
+                        website: formData.website,
+                        linkedin: formData.linkedin,
+                        overview: formData.overview,
+                        headquarters: formData.headquarters,
+                        companySize: formData.companySize === '' ? null : formData.companySize,
+                        regions: regions.map(region => ({name: region})),
+                        therapeutics: therapeutics.map(therapeutic => ({name: therapeutic})),
+                        phases: phases.length === 0 ? [] : phases,
+                        services: services.length === 0 ? [] : services
+                    }})
+                    setSubmitting(false);
+                }catch(error){
+                    console.log('Submitting of new form ',error);
+                    setError('Company Name Already Exists');
+                    setSubmitting(false);
+                }
+            }
+        }
+    }
 
-//     /*
-//         Filling in logo url
-//         - If there is a logo url stored in the database, overwrite the default logo
-//     */
-//     useEffect(() => {
-//         if(formData.logoURL){
-//             setLogo(formData.logoURL);
-//         }
-//     }, [ formData.logoURL ]);
+    const handleChange = e => {
+        setFormData({
+            ...formData,
+            [e.target.name]: e.target.value
+        })
+    }
 
-//     /*
-//         Re-direct on successful submission
-//         - When creating a new company, if an id is successfully returned, re-direct to the new company profile
-//     */
-//     useEffect(() => {
-//         if(!props.edit && newId && !addLoading){
-//             history.push(`/service-providers/${newId}`)
-//         }
-//     }, [ props.edit, newId, addLoading, history ])
+    // Handles state change for selecting specialties
+    const handleSpecialtyChange = e => {
+        const service = services.filter(service => service.name === e.target.name);
+        let specialtiesArray = [];
+        if(service[0].specialties){
+            specialtiesArray = service[0].specialties;
+        }
+        setServices([
+            ...services.filter(service => service.name !== e.target.name),
+            {
+                name: e.target.name,
+                specialties: [
+                    ...specialtiesArray,
+                    {
+                        name: e.target.value
+                    }
+                ]
+            }
+        ])
+    } 
 
-//     useEffect(() => {
-//         if(headerSuccess === true){
-//             setTimeout(() => {
-//                 setHeaderSuccess(false);
-//             }, 3000)
-//         }
-//     }, [ headerSuccess ]);
+    // Handles deleting a specialty from a service
+    const handleSpecialtyDelete = (serviceName, specialtyName) => {
+        const service = services.filter(service => service.name === serviceName);
+        setServices([
+            ...services.filter(service => service.name !== serviceName),
+            {
+                name: serviceName,
+                specialties: service[0].specialties.filter(specialty => specialty.name !== specialtyName)
+            }
+        ])
+    }
 
-//     return (
-//         <Body>
-//             <Backdrop className={classes.backdrop} open={open}>
-//                 {importWarning ? (
-//                     <WarningCard close={handleWarningClose} warning='import'/>
-//                 ) : (
-//                     <WarningCard close={handleWarningClose} warning='cancel' action={handleReDirect} data-testid='warning-card'/>
-//                 )}
-//             </Backdrop>
-//             <Backdrop className={classes.backdrop} open={companyLoading}>
-//                 <CircularProgress color="inherit" />
-//             </Backdrop>
-//             <Backdrop className={classes.backdrop} open={addLoading}>
-//                 <CircularProgress color="inherit" />
-//             </Backdrop>
-//             <HeaderError display={headerError ? 1 : 0} data-testid='header-error'>
-//                 <p>{headerErrorText}</p>
-//             </HeaderError>
-//             <SucceessMessage display={headerSuccess ? 1 : 0} data-testid='header-success'>
-//                 <p>Changes were successfully saved</p>
-//             </SucceessMessage>
-//             <div className='body'>
-//                 <div className='header-wrapper'>
-//                     {props.edit ? (
-//                         <h1>Edit company profile</h1>
-//                     ) : (
-//                         <h1>Add company profile</h1>
-//                     )}
-//                     <div className='btn-container'>
-//                         <Button onClick={handleImportWarning}>
-//                             <LinkedIn/>
-//                             <p>Import</p>
-//                         </Button>
-//                         <Button onClick={handleCancelWarning} cancel='true'>
-//                             <p>Cancel</p>
-//                         </Button>    
-//                     </div>
-//                 </div>
-//                 <div className='bar'/>
-//                 <Form>
-//                     <div className='top-row'>
-//                         <div className='side-bar'>
-//                             <img src={logo} alt='company'/>
-//                             <Button noMargin onClick={handleSubmit} id='submit-test'>
-//                                 {props.edit ? (
-//                                     <p>Save changes</p>
-//                                 ) : (
-//                                     <p>Add company</p>
-//                                 )}
-//                             </Button>
-//                         </div>
-//                         <div className='col-custom'>
-//                             <div className='input-container'>
-//                                 <label>Company Name</label>
-//                                 <input
-//                                     name='name'
-//                                     value={formData.name}
-//                                     onChange={handleUpdate}
-//                                     data-testid='name'
-//                                 />
-//                             </div>
-//                             <div className='input-container'>
-//                                 <label>Website</label>
-//                                 <input
-//                                     name='website'
-//                                     value={formData.website}
-//                                     onChange={handleUpdate}
-//                                 />
-//                             </div>
-//                             <div className='input-container'>
-//                                 <label>Company Description</label>
-//                                 <textarea
-//                                     name='overview'
-//                                     value={formData.overview}
-//                                     onChange={handleUpdate}
-//                                 />
-//                             </div>
-//                         </div>
-//                         <div className='col-custom'>
-//                             <div className='input-container' style={{
-//                                  marginBottom: linkedInError ? '18px' : '40px'
-//                             }}>
-//                                 <label>LinkedIn</label>
-//                                 <div className='linkedIn-container'>
-//                                     <input
-//                                         name='linkedin'
-//                                         value={formData.linkedin}
-//                                         onChange={handleUpdate}
-//                                         style={{ 
-//                                             borderColor: linkedInError ? 'red' : theme.colors.silver,
-//                                         }}
-//                                     />
-//                                     <p className='error'>{appError}</p>
-//                                 </div>
-//                             </div>
-//                             <div className='input-container'>
-//                                 <label>Headquarters</label>
-//                                 <input
-//                                     name='headquarters'
-//                                     value={formData.headquarters}
-//                                     onChange={handleUpdate}
-//                                 />
-//                             </div>
-//                             <div className='input-container'>
-//                                 <label>Company Size</label>
-//                                 <select 
-//                                     name='companySize' 
-//                                     value={formData.companySize}
-//                                     onChange={handleUpdate}
-//                                 >
-//                                     <option value='' defaultValue disabled hidden>Choose company size</option>
-//                                     <option value='N/A'>N/A</option>
-//                                     <option value='A'>A: Self Employed</option>
-//                                     <option value='B'>B: 1-10 Employees</option>
-//                                     <option value='C'>C: 11-50 Employees</option>
-//                                     <option value='D'>D: 51-200 Employees</option>
-//                                     <option value='E'>E: 201-500 Employees</option>
-//                                     <option value='F'>F: 501-1,000 Employees</option>
-//                                     <option value='G'>G: 1,001-5,000 Employees</option>
-//                                     <option value='H'>H: 5,001-10,000 Employees</option>
-//                                     <option value='I'>I: 10,000+ Employees</option>
-//                                 </select>
-//                             </div>
-//                         </div>
-//                     </div>
-//                     <div className='bottom-row'>
-//                         <div className='multi-container'>
-//                             <label>Regions Covered</label>
-//                             {regions && (
-//                                 <MultipleInput 
-//                                 name='regionsCovered' 
-//                                 suggestions={regions} 
-//                                 preview={formData.regionsCovered}
-//                                 handleMultiUpdate={handleMultiUpdate}
-//                             />
-//                             )}
-//                         </div>
-//                         <div className='multi-container'>
-//                             <label>Therapeutic Areas</label>
-//                             {therapeutics && (
-//                                 <MultipleInput 
-//                                 name='therapeuticAreas' 
-//                                 suggestions={therapeutics} 
-//                                 preview={formData.therapeuticAreas}
-//                                 handleMultiUpdate={handleMultiUpdate}
-//                             />
-//                             )}
-//                         </div>
-//                         <div className='multi-container'>
-//                             <label>Services</label>
-//                             {services && (
-//                                 <MultipleInput 
-//                                 name='services' 
-//                                 suggestions={services} 
-//                                 preview={formData.services}
-//                                 handleMultiUpdate={handleMultiUpdate}
-//                             />
-//                             )}
-//                         </div>
-//                         <div className='multi-container'>
-//                             <label>Specialties</label>
-//                             {specialties && (
-//                                 <MultipleInput 
-//                                 name='specialties' 
-//                                 suggestions={specialties}
-//                                 preview={formData.specialties} 
-//                                 handleMultiUpdate={handleMultiUpdate}
-//                             />
-//                             )}
-//                         </div>
-//                     </div>
-//                     <img className='background-asset' src={scienceAsset} alt=''/>
-//                 </Form>
-//             </div>
-//         </Body>
-//     );
-// }
+    // Handles state change for specialties that the user inputs
+    const handleCustomSpecialty = (serviceName, newSpecialty) => {
+        const service = services.filter(service => service.name === serviceName);
+        let specialtiesArray = [];
+        if(service[0].specialties){
+            specialtiesArray = service[0].specialties;
+        }
+        setServices([
+            ...services.filter(service => service.name !== serviceName),
+            {
+                name: serviceName,
+                specialties: [
+                    ...specialtiesArray,
+                    {
+                        name: newSpecialty
+                    }
+                ]
+            }
+        ])
+    }
+
+    // Handles state change for selecting subSpecialties
+    const handleSubSelect = (e, specialtyName) => {
+        const service = services.filter(service => service.name === e.target.name);
+
+        const oldSpecialties = service[0].specialties.filter(specialty => specialty.name !== specialtyName);
+        const updatedSpecialty = service[0].specialties.filter(specialty => specialty.name === specialtyName);
+        let oldSubSpecialties = [];
+        if(updatedSpecialty[0].sub_specialties){
+            oldSubSpecialties = updatedSpecialty[0].sub_specialties.map(subSpecialty => ({name: subSpecialty.name}));
+        }
+        setServices([
+            ...services.filter(service => service.name !== e.target.name),
+            {
+                name: e.target.name,
+                specialties: [
+                    ...oldSpecialties,
+                    {
+                        name: specialtyName,
+                        sub_specialties: [
+                            ...oldSubSpecialties,
+                            {
+                                name: e.target.value
+                            }
+                        ]
+                    }
+                ]
+            }
+        ])
+    }
+
+    // Handles deleting a sub-specialty from a specialty
+    const handleSubDelete = (serviceName, specialtyName, delSub) => {
+        const service = services.filter(service => service.name === serviceName);
+        const oldSpecialties = service[0].specialties.filter(specialty => specialty.name !== specialtyName);
+        const updatedSpecialty = service[0].specialties.filter(specialty => specialty.name === specialtyName);
+        console.log(oldSpecialties)
+        setServices([
+            ...services.filter(service => service.name !== serviceName),
+            {
+                name: serviceName,
+                specialties: [
+                    ...oldSpecialties,
+                    {
+                        name: specialtyName,
+                        sub_specialties: updatedSpecialty[0].sub_specialties.filter(subSpecialty => subSpecialty.name !== delSub)
+                    }
+                ]
+            }
+        ])
+    }
+
+    // Handles state change for subSpecialties the user inputs
+    const handleCustomSub = (serviceName, specialtyName, newSub) => {
+        const service = services.filter(service => service.name === serviceName);
+        const oldSpecialties = service[0].specialties.filter(specialty => specialty.name !== specialtyName);
+        const updatedSpecialty = service[0].specialties.filter(specialty => specialty.name === specialtyName);
+        let oldSubSpecialties = []
+        if(updatedSpecialty[0].sub_specialties){
+            oldSubSpecialties = updatedSpecialty[0].sub_specialties.map(subSpecialty => ({name: subSpecialty.name}))
+        }
+        setServices([
+            ...services.filter(service => service.name !== serviceName),
+            {
+                name: serviceName,
+                specialties: [
+                    ...oldSpecialties,
+                    {
+                        name: specialtyName,
+                        sub_specialties: [
+                            ...oldSubSpecialties,
+                            {
+                                name: newSub
+                            }
+                        ]
+                    }
+                ],
+            }
+        ])
+    }
+
+    useEffect(() => {
+        if(regionsAll){
+            setAll('regions');
+        }else{
+            clearAll('regions');
+        }
+        if(therapeuticsAll){
+            setAll('therapeutics');
+        }else{
+            clearAll('therapeutics');
+        }
+        if(phasesAll){
+            setAll('phases');
+        }else{
+            clearAll('phases');
+        }
+    }, [ therapeuticsAll, regionsAll, phasesAll ]);
+
+    // Error handling forr fetching data by id
+    useEffect(() => {
+        if(companyError){
+            console.log('Company fetching error ', companyError);
+        }
+    }, [ companyError ])
+
+    useEffect(() => {
+        if(edit){
+            if(companyData){
+                const regionsMapped = companyData.company.regions.map(region => region.name);
+                const therapeuticsMapped = companyData.company.therapeutics.map(therapeutic => therapeutic.name);
+                setFormData({
+                    name: companyData.company.name,
+                    email: companyData.company.email,
+                    logoURL: companyData.company.logoURL,
+                    website: companyData.company.website,
+                    linkedin: companyData.company.linkedin,
+                    overview: companyData.company.overview,
+                    headquarters: companyData.company.headquarters,
+                    companySize: companyData.company.companySize ? companyData.company.companySize : '',
+                })
+                setRegions(regionsMapped);
+                setTherapeutics(therapeuticsMapped)
+            }
+        }
+    }, [ edit, companyData ])
+    
+    return (
+        <Body>
+            {confirmCancel && (
+                <Backdrop className={classes.backdrop} open={confirmCancel}>
+                    <WarningCard handleCancel={handleCancel} handleReDirect={handleReDirect}/>
+                </Backdrop>
+            )}
+            {submitting && (
+                <Backdrop className={classes.backdrop} open={submitting}>
+                    <CircularProgress color='inherit'/>
+                </Backdrop>
+            )}
+            {loading && (
+                <Backdrop className={classes.backdrop} open={submitting}>
+                    <CircularProgress color='inherit'/>
+                </Backdrop>
+            )}
+            <header>
+                <div className='header-content'>
+                    <h2>Create Company Profile</h2>
+                    <div className='btn-container' onClick={handleCancel}>
+                        <Arrow/>
+                        <p className='grey'>Back</p>
+                    </div>
+                </div>
+            </header>
+            {error.length > 1 && (
+                <div className='error'>
+                    <p>{error}</p>
+                </div>
+            )}
+            <div className='form-wrapper'>
+                <div className='import-container'>
+                    <h3>LinkedIn Import</h3>
+                    <p className='text'>Would you like to import your LinkedIn data to fill out your basic information?</p>
+                    <div className='warning-container'>
+                        <WarningIcon/>
+                        <p className='text'>Warning: All current data on this page will be overwritten.</p>
+                    </div>
+                    <Button color='scienceBlue'><p>Import</p></Button>
+                </div>
+                <Basic>
+                    <h3>Basic Information</h3>
+                    <div className='form-wrapper'>
+                        <div className='form'>
+                            <div className='row'>
+                                <div className='input-box'>
+                                    <label>Company Name*</label>
+                                    <input
+                                        name='name'
+                                        onChange={handleChange}
+                                        value={formData.name}
+                                    />
+                                </div>
+                                <div className='input-box'>
+                                    <label>RFP Email</label>
+                                    <input
+                                        name='email'
+                                        onChange={handleChange}
+                                        value={formData.email}
+                                    />
+                                </div>
+                            </div>
+                            <div className='row'>
+                                <div className='input-box'>
+                                    <label>LinkedIn</label>
+                                    <input
+                                        name='linkedin'
+                                        onChange={handleChange}
+                                        value={formData.linkedin}
+                                    />
+                                </div>
+                                <div className='input-box'>
+                                    <label>Website</label>
+                                    <input
+                                        name='website'
+                                        onChange={handleChange}
+                                        value={formData.website}
+                                    />
+                                </div>
+                            </div>
+                            <div className='row'>
+                                <div className='input-box'>
+                                    <label>Headquarters</label>
+                                    <input
+                                        name='headquarters'
+                                        onChange={handleChange}
+                                        value={formData.headquarters}
+                                    />
+                                </div>
+                                <div className='input-box'>
+                                    <label>Company Size</label>
+                                    <FormControl className={classes.formControl}>
+                                        <Select
+                                            labelId="demo-simple-select-label"
+                                            id="demo-simple-select"
+                                            value={formData.companySize}
+                                            onChange={handleChange}
+                                            name='companySize'
+                                        >
+                                            <MenuItem value=''>N/A</MenuItem>
+                                            <MenuItem value='A'>Self Employed</MenuItem>
+                                            <MenuItem value='B'>1-10 Employees</MenuItem>
+                                            <MenuItem value='C'>11-50 Employees</MenuItem>
+                                            <MenuItem value='D'>51-200 Employees</MenuItem>
+                                            <MenuItem value='E'>201-500 Employees</MenuItem>
+                                            <MenuItem value='F'>501-1,000 Employees</MenuItem>
+                                            <MenuItem value='G'>1,001-5,000 Employees</MenuItem>
+                                            <MenuItem value='H'>5,001-10,000 Employees</MenuItem>
+                                            <MenuItem value='I'>10,000+ Employees</MenuItem>
+                                        </Select>
+                                    </FormControl>
+                                </div>
+                            </div>
+                            <div className='row'>
+                                <div className='input-box'>
+                                    <label>Regions Covered</label>
+                                    {regionsData && (
+                                        <FormControl className={classes.formControl}>
+                                            <Select
+                                                labelId="demo-mutiple-checkbox-label"
+                                                id="demo-mutiple-checkbox"
+                                                multiple
+                                                value={regions}
+                                                onChange={handleMultiple}
+                                                input={<Input />}
+                                                renderValue={(selected) => selected.join(', ')}
+                                                MenuProps={MenuProps}
+                                                name='regions'
+                                                placeholder='Select Regions'
+                                            >
+                                            <MenuItem key='All' value='All' onClick={toggleAllRegions}>
+                                                <Checkbox checked={regionsAll}/>
+                                                <ListItemText primary='All'/>
+                                            </MenuItem>
+                                            {regionsData.regions.map(region => (
+                                                <MenuItem key={region.name} value={region.name}>
+                                                    <Checkbox checked={regions.indexOf(region.name) > -1} />
+                                                    <ListItemText primary={region.name} />
+                                                </MenuItem>
+                                            ))}
+                                            </Select>
+                                        </FormControl>
+                                    )}
+                                </div>
+                                <div className='input-box'>
+                                    <label>Therapeutic Areas</label>
+                                    {therapeuticsData && (
+                                        <FormControl className={classes.formControl}>
+                                            <Select
+                                                labelId="demo-mutiple-checkbox-label"
+                                                id="demo-mutiple-checkbox"
+                                                multiple
+                                                value={therapeutics}
+                                                onChange={handleMultiple}
+                                                input={<Input />}
+                                                renderValue={(selected) => selected.join(', ')}
+                                                MenuProps={MenuProps}
+                                                name='therapeutics'
+                                            >
+                                            <MenuItem key='All' value='All' onClick={toggleAllTherapeutics}>
+                                                <Checkbox checked={therapeuticsAll}/>
+                                                <ListItemText primary='All'/>
+                                            </MenuItem>
+                                            {therapeuticsData.therapeutics.map(therapeutic => (
+                                                <MenuItem key={therapeutic.name} value={therapeutic.name}>
+                                                    <Checkbox checked={therapeutics.indexOf(therapeutic.name) > -1} />
+                                                    <ListItemText primary={therapeutic.name} />
+                                                </MenuItem>
+                                            ))}
+                                            </Select>
+                                        </FormControl>
+                                    )}
+                                </div>
+                            </div>
+                            <div className='row'>
+                                <div className='input-box'>
+                                    <label>Clinical Trial Phases</label>
+                                    <FormControl className={classes.formControl}>
+                                        <Select
+                                            labelId="demo-mutiple-checkbox-label"
+                                            id="demo-mutiple-checkbox"
+                                            multiple
+                                            value={phases}
+                                            onChange={handleMultiple}
+                                            input={<Input />}
+                                            renderValue={(selected) => selected.join(', ')}
+                                            MenuProps={MenuProps}
+                                            name='phases'
+                                        >
+                                        <MenuItem key='All' value='All' onClick={toggleAllPhases}>
+                                            <Checkbox checked={phasesAll}/>
+                                            <ListItemText primary='All'/>
+                                        </MenuItem>
+                                        {phaseOptions.map(phase => (
+                                            <MenuItem key={phase} value={phase}>
+                                                <Checkbox checked={phases.indexOf(phase) > -1} />
+                                                <ListItemText primary={phase} />
+                                            </MenuItem>
+                                        ))}
+                                        </Select>
+                                    </FormControl>
+                                </div>
+                                <div className='input-box'>
+                                    <label>Company Overview</label>
+                                    <textarea
+                                        name='overview'
+                                        value={formData.overview}
+                                        onChange={handleChange}
+                                    />
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </Basic>
+                <Services>
+                    <h3>Services</h3>
+                    <div className='service-container'>
+                        <div className='wrapper'>
+                            <div className='container-col'>
+                                <label>Add Services</label>
+                                {servicesData && (
+                                    <FormControl className={classes.formControl}>
+                                        <Select
+                                            labelId="demo-mutiple-checkbox-label"
+                                            id="demo-mutiple-checkbox"
+                                            multiple
+                                            value={services.map(service => service.name)}
+                                            onChange={handleMultiple}
+                                            input={<Input />}
+                                            renderValue={(selected) => selected.join(', ')}
+                                            MenuProps={MenuProps}
+                                            name='services'
+                                        >
+                                        {servicesData.serviceItems.map(service => (
+                                            <MenuItem key={service.name} value={service.name}>
+                                                <Checkbox checked={services.map(service => service.name).indexOf(service.name) > -1} />
+                                                <ListItemText primary={service.name} />
+                                            </MenuItem>
+                                        ))}
+                                        </Select>
+                                    </FormControl>
+                                )}
+                                {specialtyData && services.map(service => {
+                                    return <Service 
+                                                service={service} 
+                                                specialtyData={specialtyData} 
+                                                handleSpecialtyChange={handleSpecialtyChange}
+                                                handleSpecialtyDelete={handleSpecialtyDelete}
+                                                handleCustomSpecialty={handleCustomSpecialty}
+                                                handleCustomSub={handleCustomSub}
+                                                handleSubSelect={handleSubSelect}
+                                                handleSubDelete={handleSubDelete}
+                                                key={service.name}
+                                            />
+                                })}
+                                
+                            </div>     
+                            <div className='container-col'>
+                                <p>Overview</p>
+                            </div>             
+                        </div>
+                    </div>
+                </Services>
+                <div className='btn-wrapper'>
+                    <div className='btn-container'>
+                        <Button color='sun' onClick={handleCancel}><p>Cancel</p></Button>
+                            <Button color='scienceBlue' marginLeft='10px' onClick={handleSubmit}><p>{edit ? 'Save' : 'Submit'}</p></Button>
+                    </div>
+                </div>
+            </div>
+        </Body>
+    );
+}
